pyproject.toml
[project]
name = "dbr-serverless-connect"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Daniel Steman", email = "daniel-steman@live.nl" }
]
dependencies = [
    "databricks-connect>=15.3.1",
    "pyyaml>=6.0.1",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["localbricks"]
